+++ src/Units/Diplomat.cs
@@ -112,7 +112,7 @@ namespace CivOne.Units
 		}
 
 		internal void KeepMoving(IUnit unit) => MovementTo(unit.X - X, unit.Y - Y);
-		
+
 		public Diplomat() : base(3, 0, 0, 2)
 		{
 			Type = UnitType.Diplomat;
@@ -120,7 +120,14 @@ namespace CivOne.Units
 			RequiredTech = new Writing();
 			ObsoleteTech = null;
 			SetIcon('C', 1, 0);
-            Role = UnitRole.Civilian;
+			Role = UnitRole.Civilian;
+			// CW: Hot fix to disband AI Diplomat units, when WorkProgress hits 0
+			// WorkProgress will be reset if Diplomat finds "friendly" unit.
+			// 10 = 5 turns of 2 moving points + random max 10 more turns [from 5 up to 15 turns]
+			// Resets if friendly unit found.
+			// Duplicate code in AI.Barbarians.cs
+			// WARNING: This is not from the original code!
+			WorkProgress = (byte)(10 + Common.Random.Next(0, 20));
         }
 	}
 }